requirements.txt
tqdm==4.67.1
urllib3==2.4.0
snoop==0.6.0
websockets==15.0.1
ffmpeg-python==0.2.0
pyodbc==5.2.0
python_dotenv==1.1.0
psutil>=5.8.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytest==8.4.2
requests
pika
